added crit and maxhp buffs

+++ Assets/Scripts/NPC/NPCHero.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "NPCHero_", menuName = "Game/NPC/Hero")]
 public class NPCHero : NPCData
 {
-    public int Strength;
-    public int MaxHp;
-    public int Hp;
+    public float Strength;
+    public float MaxHp;
+    public float Hp;
 }
